refactor(ai-agent): replace broadcast consumer with domain-event mpsc (#7b-4)
The persistence consumer now subscribes to `mpsc::Receiver<AcpSessionEvent>`
emitted by the AcpSession aggregate, instead of reverse-engineering user
intent from CLI observation events on the broadcast bus.

Key changes:
- Add `domain_event_tx` field to AcpAgentManager; `commit_session_changes()`
  drains the aggregate's pending events and forwards them via mpsc.
- New `manager/acp/session_sync.rs` (`AcpSessionSyncService`) replaces the
  old `acp_agent_service.rs`. Consumer only persists DesiredModeChanged /
  DesiredConfigChanged / ObservedModelSynced — no more `currentModeId`
  reverse-engineering.
- CLI detect helpers moved to `protocol/cli_detect.rs`.
- Delete `acp_agent_service.rs` (642 lines).
