+++ crates/aionui-ai-agent/src/agent_routes.rs
@@ -9,8 +9,8 @@ use aionui_api_types::{AgentMetadata, ApiResponse, TestCustomAgentRequest, TestC
 use aionui_auth::CurrentUser;
 use aionui_common::AppError;
 
-use crate::acp_agent_service;
 use crate::agent_registry::AgentRegistry;
+use crate::protocol::cli_detect;
 
 #[derive(Clone)]
 pub struct AgentRouterState {
@@ -47,6 +47,6 @@ async fn test_custom_agent(
     body: Result<Json<TestCustomAgentRequest>, JsonRejection>,
 ) -> Result<Json<ApiResponse<TestCustomAgentResponse>>, AppError> {
     let Json(req) = body.map_err(|e| AppError::BadRequest(e.to_string()))?;
-    let result = acp_agent_service::test_custom_agent(&req.command, &req.acp_args, &req.env)?;
+    let result = cli_detect::test_custom_agent(&req.command, &req.acp_args, &req.env)?;
     Ok(Json(ApiResponse::ok(result)))
 }
