+++ crates/aionui-ai-agent/src/manager/acp/mod.rs
@@ -1,7 +1,9 @@
 pub mod events;
 pub mod reconcile;
 pub mod session;
+pub mod session_sync;
 
 pub use events::AcpSessionEvent;
 pub use reconcile::ReconcileAction;
 pub use session::{AcpSession, PersistedSessionState};
+pub use session_sync::AcpSessionSyncService;
