+++ crates/aionui-ai-agent/src/protocol/mod.rs
@@ -1,1 +1,2 @@
+pub mod cli_detect;
 pub mod events;
