+++ crates/aionui-ai-agent/tests/factory_provider_integration.rs
@@ -1,7 +1,7 @@
 use std::path::PathBuf;
 use std::sync::Arc;
 
-use aionui_ai_agent::acp_agent_service::AcpAgentService;
+use aionui_ai_agent::AcpSessionSyncService;
 use aionui_ai_agent::agent_registry::AgentRegistry;
 use aionui_ai_agent::factory::{AgentFactoryDeps, build_agent_factory};
 use aionui_ai_agent::skill_manager::AcpSkillManager;
@@ -20,7 +20,7 @@ async fn setup() -> (
     Arc<dyn IProviderRepository>,
     Arc<SqliteRemoteAgentRepository>,
     Arc<AgentRegistry>,
-    Arc<AcpAgentService>,
+    Arc<AcpSessionSyncService>,
 ) {
     let db = init_database_memory().await.unwrap();
     let pool = db.pool().clone();
@@ -30,7 +30,7 @@ async fn setup() -> (
     let registry = AgentRegistry::new(metadata_repo);
     registry.hydrate().await.unwrap();
     let session_repo: Arc<dyn IAcpSessionRepository> = Arc::new(SqliteAcpSessionRepository::new(pool));
-    let acp_agent_service = AcpAgentService::new(session_repo);
+    let acp_agent_service = AcpSessionSyncService::new(session_repo);
     (provider_repo, remote_agent_repo, registry, acp_agent_service)
 }
 
@@ -60,7 +60,7 @@ fn make_factory(
     provider_repo: Arc<dyn IProviderRepository>,
     remote_agent_repo: Arc<SqliteRemoteAgentRepository>,
     agent_registry: Arc<AgentRegistry>,
-    acp_agent_service: Arc<AcpAgentService>,
+    acp_agent_service: Arc<AcpSessionSyncService>,
 ) -> aionui_ai_agent::AgentFactory {
     let tmp = tempfile::TempDir::new().unwrap();
     let skill_paths = Arc::new(aionui_extension::resolve_skill_paths(tmp.path(), tmp.path()));
